[llvm-dec] Don't use cl::opt<uint64_t> (PR19665).
It breaks, but only on linux+gcc, which I guess has a typedef on
something other than unsigned long long.

+++ tools/llvm-dec/llvm-dec.cpp
@@ -40,7 +40,8 @@ static cl::opt<std::string>
 TripleName("triple", cl::desc("Target triple to disassemble for, "
                               "see -version for available targets"));
 
-static cl::opt<uint64_t>
+// cl::opt<uint64_t> isn't currently supported (PR19665).
+static cl::opt<unsigned long long>
 TranslationEntrypoint("entrypoint",
                       cl::desc("Address to start translating from "
                                "(default = object entrypoint)"));
